--HG--
extra : convert_revision : 8a558785c64b7c33e64523d3d887ea6e760c3d2b

+++ dev/tsunami_cchip.cc
@@ -176,7 +176,7 @@ TsunamiCChip::read(MemReqPtr &req, uint8_t *data)
           if (regnum == TSDEV_CC_DRIR) {
               warn("accessing DRIR with 32 bit read, "
                    "hopefully your just reading this for timing");
-              *(uint64_t*)data = drir;
+              *(uint32_t*)data = drir;
           } else
               panic("invalid access size(?) for tsunami register!\n");
           return No_Fault;
